[refactor] #53 - 리마인드 모달 주기 타입별 컴포넌트 분리
